requirements.txt
opencv-python-headless
Pillow
pytesseract
tk
numpy
tensorflow
tkvideoplayer